backend/requirements.txt
# Core Framework
fastapi==0.111.*
uvicorn[standard]==0.29.*
python-socketio==5.*

# Database & ORM
sqlmodel==0.0.16
psycopg2-binary==2.9.*
alembic==1.13.*

# Caching & Message Queue
redis==5.*
celery[redis]==5.*

# Trading & Market Data
Dhan-Tradehull==3.*

# Data Processing & ML
pandas==2.2.*
numpy==2.2.*
scikit-learn==1.5.*

# Async & HTTP
aiohttp==3.9.*
httpx==0.27.*

# Utilities
pydantic==2.6.*
pydantic-settings==2.2.*
python-dotenv==1.0.*
loguru==0.7.*
pytz==2024.*
psutil==5.*

# Development & Testing
pytest==8.1.*
pytest-asyncio==0.23.*
respx==0.21.*
ruff==0.3.*
mypy==1.9.*

# Security
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.* 